confirmar em master

+++ src/multistar.jl
@@ -6,7 +6,7 @@ function multistar_inequalities!(
     edge_tail::Vector{Int}, 
     edge_head::Vector{Int}, 
     edge_x::Vector{Float64};
-    max_n_cuts = 100
+    max_n_cuts = 100::Int
 )
     n_customers, _demand, n_edges, _edge_tail, _edge_head, _edge_x = input_conversion(demand, edge_tail, edge_head, edge_x)
